player classes organized without dupplicate code
